Going to attempt to make Git recognize that these are the same file
